fix: コントローラーを移行 railsの思想に合っていないため、リソースに属したコンローラーを作成
